Copy signalflow dir from signalfx-python
This change copies (`cp -r`) the directory at
https://github.com/signalfx/signalfx-python/tree/master/signalfx/signalflow
and a couple of other files to this new repo.

This change doesn't actually compile because it references code that is not
present after the copy, but I want to keep it as-is for this "copy" commit
and will fix/improve things in later ones.

Changes attributed to signalfx@signalfx.com